requirements.txt
pyTelegramBotAPI
python-dotenv
requests
openai